feat: `MerkleVCA` contract (#58)
* feat: MerkleVCA contract

refactor: separate out parameters struct for Merkle Instant and Merkle Lockups
refactor: refactor shape and camapign name as bytes32 in Merkle campaign constructor arguments

chore: contract inheritance

test: deploy factory locally for fork tests

feat: merkle vca contract

feat: add CreateMerkleVCA run script
refactor: rename schedule to _vestingSchedule in MerkleVCA
feat: add SablierMerkleVCA in prepare-artifacts bash script
refactor: replace internal variables with private in non-abstract contracts

refactor: 7 days claim window post vesting unlock
fix: set sacrificedAmount to 0 in refundSacrificedAmount function
refactor: error name

feat: revert claim if vesting start time is in future
feat: revert createMerkleVCA if either of vesting times is zero

test: integration tests for createMerkleVCA

fix: calculate claimable amount only if blocktime is in the vesting range
test: add merkleVCA integration tests

test: integration tests

test: fork tests for merkleVCA

test: remove Vars struct from merkle instant contructor test

refactore: rename sacrifiedAmount to forgoneAmount

* chore: use minimum fee

* refactor: move input checks to merkleVCA constructor


* refactor: pre declare the blockTimestamp in _claim

chore: use named args in transfer function
chore: correct prepare artifacts interface

* andrei's feedback


* feat: revert vca if expiry is zero

* docs: explanatory comment about cast

test: rename dir to tokens
small polishes

* test: fix fork tests

* docs: correct notice natspec

---------

+++ tests/fork/merkle-campaign/MerkleLL.t.sol
@@ -241,7 +241,7 @@ abstract contract MerkleLL_Fork_Test is Fork_Test {
 
         if (params.expiration > 0) {
             vars.clawbackAmount = uint128(FORK_TOKEN.balanceOf(address(vars.merkleLL)));
-            vm.warp({ newTimestamp: uint256(params.expiration) + 1 seconds });
+            vm.warp({ newTimestamp: params.expiration + 1 seconds });
 
             expectCallToTransfer({ token: FORK_TOKEN, to: params.campaignOwner, value: vars.clawbackAmount });
             vm.expectEmit({ emitter: address(vars.merkleLL) });
